Fix teacher logic

+++ project4/src/main/java/com/epam/jwd/controller/command/DeleteCourseCommand.java
@@ -11,7 +11,9 @@ import com.epam.jwd.service.dto.userdto.UserDto;
 import com.epam.jwd.service.exception.ServiceException;
 import com.epam.jwd.service.impl.CourseService;
 import com.epam.jwd.service.impl.ReviewService;
-import com.epam.jwd.service.impl.UserService;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
 
 import java.util.ArrayList;
 import java.util.List;
@@ -19,17 +21,23 @@ import java.util.List;
 
 public class DeleteCourseCommand implements Command {
 
+    private static final Logger LOGGER = LogManager.getLogger(DeleteCourseCommand.class);
+
     private static final Command INSTANCE = new DeleteCourseCommand();
+
     private static final String ERROR_SESSION_COLLECTION_ATTRIBUTE = "errorName";
     private static final String CANNOT_FIND_COURSE_MESSAGE = "I can't find this course";
     private static final String CANNOT_DELETE_COURSE_MESSAGE = "I can't delete all courses by course_id";
+
     private static final String ERROR_COURSE_COMMAND = "/controller?command=SHOW_ERROR_PAGE_COMMAND";
     private static final String DELETE_COURSE_COMMAND = "/controller?command=SHOW_DELETE_COURSE_PAGE_COMMAND";
     private static final String TEACHER_RESULT_COMMAND = "/controller?command=SHOW_TEACHER_PAGE_COMMAND";
+
     private static final String USER_COURSE_SESSION_COLLECTION_ATTRIBUTE = "userCourse";
+
     private final Service<CourseDto, Integer> courseService = new CourseService();
     private final Service<ReviewDto, Integer> reviewService = new ReviewService();
-    private final Service<UserDto, Integer> userService = new UserService();
+
 
     public static Command getInstance() {
         return INSTANCE;
@@ -78,47 +86,61 @@ public class DeleteCourseCommand implements Command {
         }
     };
 
+
     @Override
     public ResponseContext execute(RequestContext requestContext) {
+
         String btnDeleteCourse = requestContext.getParameterFromJSP("btnDeleteCourse");
         String btnGetBack = requestContext.getParameterFromJSP("btnGetBack");
+
+        if(btnGetBack !=null){
+            return TEACHER_RESULT_CONTEXT;
+        }
 
         UserDto userDto = (UserDto) requestContext.getAttributeFromSession("currentUser");
 
         if (btnDeleteCourse != null){
-            String id = requestContext.getParameterFromJSP("lblDelete");
-            int course_id = Integer.parseInt(id);
-            CourseDto courseDtoForDelete = courseService.getById(course_id);
-            if (courseDtoForDelete == null){
+
+            String name = requestContext.getParameterFromJSP("Course_name");
+
+            List<CourseDto> courseList = ((CourseService) courseService).filterCourse(name);
+
+            if (courseList.size() == 0){
+                LOGGER.error(CANNOT_FIND_COURSE_MESSAGE);
                 requestContext.addAttributeToSession(ERROR_SESSION_COLLECTION_ATTRIBUTE,CANNOT_FIND_COURSE_MESSAGE);
                 return ERROR_PAGE_CONTEXT;
             }
-            //delete all review where course_id
+
+            CourseDto courseDtoForDelete = courseList.get(0);
+
             List<ReviewDto> listOfThisCourseReview = new ArrayList<>();
             try {
-                listOfThisCourseReview = ((ReviewService)reviewService).findReviewByCourseId(course_id);
+                listOfThisCourseReview = ((ReviewService)reviewService).findReviewByCourseId(courseDtoForDelete.getId());
             }catch (DAOException exception){
-                //log
+                LOGGER.error(exception.getMessage());
             }
             if (listOfThisCourseReview.size() != 0){
                 for (ReviewDto review:
                         listOfThisCourseReview) {
                     try {
                         reviewService.delete(review);
-                    } catch (ServiceException e) {
-                        //log
+                    } catch (ServiceException exception) {
+                        LOGGER.error(exception.getMessage());
                     }
                 }
             }
-            Boolean result = ((CourseService)courseService).deleteAllCourseInUSERHAsCourse(course_id);
+            Boolean result = ((CourseService)courseService).deleteAllCourseInUSERHAsCourse(courseDtoForDelete.getId());
+
             if (!result) {
+                LOGGER.error(CANNOT_FIND_COURSE_MESSAGE);
                 requestContext.addAttributeToSession(ERROR_SESSION_COLLECTION_ATTRIBUTE, CANNOT_DELETE_COURSE_MESSAGE);
                 return ERROR_PAGE_CONTEXT;
             }
-                CourseDto courseForDelete = courseService.getById(course_id);
+                CourseDto courseForDelete = courseService.getById(courseDtoForDelete.getId());
                 try{
                     courseService.delete(courseForDelete);
                 }catch (Exception exception){
+                    LOGGER.error(exception.getMessage());
                     requestContext.addAttributeToSession(ERROR_SESSION_COLLECTION_ATTRIBUTE, exception.getMessage());
                     return ERROR_PAGE_CONTEXT;
                 }
@@ -126,8 +148,6 @@ public class DeleteCourseCommand implements Command {
             List<CourseDto> coursesAfterDelete = ((CourseService) courseService).getUserAvailableCourses(userDto.getFirst_name(),userDto.getLast_name());
             requestContext.addAttributeToSession(USER_COURSE_SESSION_COLLECTION_ATTRIBUTE, coursesAfterDelete);
             return REFRESH_PAGE_CONTEXT;
-        }else if(btnGetBack !=null){
-            return TEACHER_RESULT_CONTEXT;
         }
         return DefaultCommand.getInstance().execute(requestContext);
     }
